# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Meme Coin Scalping Bot API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type BotStatusMode = (typeof BotStatusMode)[keyof typeof BotStatusMode];

export const BotStatusMode = {
  live: "live",
  paper: "paper",
} as const;

export interface BotStatus {
  running: boolean;
  mode: BotStatusMode;
  /** @nullable */
  startedAt?: string | null;
  dailyLossHit: boolean;
  /** @nullable */
  cooldownUntil?: string | null;
  activePositions: number;
  totalTradesDay: number;
  winRateDay: number;
  /** @nullable */
  walletAddress?: string | null;
}

export type PositionStatus =
  (typeof PositionStatus)[keyof typeof PositionStatus];

export const PositionStatus = {
  open: "open",
  closing: "closing",
} as const;

export interface Position {
  id: number;
  tokenAddress: string;
  tokenSymbol: string;
  tokenName: string;
  entryPrice: number;
  currentPrice: number;
  amountEth: number;
  amountTokens: number;
  profitPercent: number;
  profitEth: number;
  entryTime: string;
  holdSeconds: number;
  tp1Hit: boolean;
  tp2Hit: boolean;
  trailingStopActive: boolean;
  safetyScore: number;
  liquidityUsd: number;
  status: PositionStatus;
}

export type TradeExitReason =
  (typeof TradeExitReason)[keyof typeof TradeExitReason];

export const TradeExitReason = {
  tp1: "tp1",
  tp2: "tp2",
  tp3: "tp3",
  stop_loss: "stop_loss",
  trailing_stop: "trailing_stop",
  max_hold: "max_hold",
  manual: "manual",
  emergency: "emergency",
} as const;

export interface Trade {
  id: number;
  tokenAddress: string;
  tokenSymbol: string;
  tokenName: string;
  entryPrice: number;
  exitPrice: number;
  amountEth: number;
  profitPercent: number;
  profitEth: number;
  entryTime: string;
  exitTime: string;
  holdSeconds: number;
  exitReason: TradeExitReason;
  /** @nullable */
  txHash?: string | null;
}

export interface TradingStats {
  todayPnlEth: number;
  todayPnlIdr: number;
  totalTradesDay: number;
  winningTradesDay: number;
  losingTradesDay: number;
  winRateDay: number;
  avgHoldSeconds: number;
  avgProfitPercent: number;
  largestWin: number;
  largestLoss: number;
  totalPnlAllTime: number;
  ethPriceUsd: number;
  capitalEth: number;
  capitalIdr: number;
}

export type ScalpingConfigRiskLevel =
  (typeof ScalpingConfigRiskLevel)[keyof typeof ScalpingConfigRiskLevel];

export const ScalpingConfigRiskLevel = {
  conservative: "conservative",
  moderate: "moderate",
  aggressive: "aggressive",
} as const;

export type ScalpingConfigMode =
  (typeof ScalpingConfigMode)[keyof typeof ScalpingConfigMode];

export const ScalpingConfigMode = {
  live: "live",
  paper: "paper",
} as const;

export interface ScalpingConfig {
  totalCapitalEth: number;
  maxTradeAmountEth: number;
  maxConcurrentPositions: number;
  min5mVolumeUsd: number;
  minLiquidityUsd: number;
  maxTokenAgeMinutes: number;
  minMomentumPercent: number;
  tp1Percent: number;
  tp2Percent: number;
  tp3Percent: number;
  stopLossPercent: number;
  trailingStopActivatePercent: number;
  trailingStopDistancePercent: number;
  maxHoldMinutes: number;
  maxDailyLossEth: number;
  maxSlippagePercent: number;
  minSafetyScore: number;
  scanIntervalSeconds: number;
  riskLevel: ScalpingConfigRiskLevel;
  mode: ScalpingConfigMode;
}

export type ScalpingConfigUpdateRiskLevel =
  (typeof ScalpingConfigUpdateRiskLevel)[keyof typeof ScalpingConfigUpdateRiskLevel];

export const ScalpingConfigUpdateRiskLevel = {
  conservative: "conservative",
  moderate: "moderate",
  aggressive: "aggressive",
} as const;

export type ScalpingConfigUpdateMode =
  (typeof ScalpingConfigUpdateMode)[keyof typeof ScalpingConfigUpdateMode];

export const ScalpingConfigUpdateMode = {
  live: "live",
  paper: "paper",
} as const;

export interface ScalpingConfigUpdate {
  totalCapitalEth?: number;
  maxTradeAmountEth?: number;
  maxConcurrentPositions?: number;
  min5mVolumeUsd?: number;
  minLiquidityUsd?: number;
  maxTokenAgeMinutes?: number;
  minMomentumPercent?: number;
  tp1Percent?: number;
  tp2Percent?: number;
  tp3Percent?: number;
  stopLossPercent?: number;
  trailingStopActivatePercent?: number;
  trailingStopDistancePercent?: number;
  maxHoldMinutes?: number;
  maxDailyLossEth?: number;
  maxSlippagePercent?: number;
  minSafetyScore?: number;
  scanIntervalSeconds?: number;
  riskLevel?: ScalpingConfigUpdateRiskLevel;
  mode?: ScalpingConfigUpdateMode;
}

export interface WalletBalance {
  /** @nullable */
  address?: string | null;
  ethBalance: number;
  ethBalanceIdr: number;
  ethPriceUsd: number;
  usdToIdr: number;
  network: string;
  /** @nullable */
  blockNumber?: number | null;
}

export interface ScannedToken {
  address: string;
  symbol: string;
  name: string;
  priceUsd: number;
  priceChangePercent5m: number;
  priceChangePercent1h: number;
  volumeUsd5m: number;
  liquidityUsd: number;
  /** @nullable */
  marketCapUsd?: number | null;
  ageMinutes: number;
  /** @nullable */
  holderCount?: number | null;
  safetyScore: number;
  passedFilters: boolean;
  scannedAt: string;
  /** @nullable */
  dexUrl?: string | null;
}

export type BotLogLevel = (typeof BotLogLevel)[keyof typeof BotLogLevel];

export const BotLogLevel = {
  info: "info",
  buy: "buy",
  sell: "sell",
  stop_loss: "stop_loss",
  warn: "warn",
  error: "error",
} as const;

export interface BotLog {
  id: number;
  level: BotLogLevel;
  message: string;
  /** @nullable */
  tokenSymbol?: string | null;
  /** @nullable */
  data?: string | null;
  timestamp: string;
}

export interface ClosePositionInput {
  /**
   * @minimum 1
   * @maximum 100
   */
  percent: number;
}

export type GetTradesParams = {
  filter?: GetTradesFilter;
  limit?: number;
};

export type GetTradesFilter =
  (typeof GetTradesFilter)[keyof typeof GetTradesFilter];

export const GetTradesFilter = {
  today: "today",
  week: "week",
  all: "all",
} as const;

export type GetScannedTokensParams = {
  limit?: number;
};

export type GetLogsParams = {
  limit?: number;
};