# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Meme Coin Scalping Bot API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /positions/{address}/close:
    post:
      operationId: closePosition
      tags: [positions]
      summary: Manually close a position
      parameters:
        - name: address
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClosePositionInput"
      responses:
        "200":
          description: Position closed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trade"
  /trades:
    get:
      operationId: getTrades
      tags: [trades]
      summary: Get trade history
      parameters:
        - name: filter
          in: query
          schema:
            type: string
            enum: [today, week, all]
            default: today
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
      responses:
        "200":
          description: Trade history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Trade"
...
components:
  schemas:
...
    Trade:
      type: object
      properties:
        id:
          type: integer
        tokenAddress:
          type: string
        tokenSymbol:
          type: string
        tokenName:
          type: string
        entryPrice:
          type: number
        exitPrice:
          type: number
        amountEth:
          type: number
        profitPercent:
          type: number
        profitEth:
          type: number
        entryTime:
          type: string
        exitTime:
          type: string
        holdSeconds:
          type: integer
        exitReason:
          type: string
          enum: [tp1, tp2, tp3, stop_loss, trailing_stop, max_hold, manual, emergency]
        txHash:
          type: ["string", "null"]
      required:
        - id
        - tokenAddress
        - tokenSymbol
        - tokenName
        - entryPrice
        - exitPrice
        - amountEth
        - profitPercent
        - profitEth
        - entryTime
        - exitTime
        - holdSeconds
        - exitReason
...
    ClosePositionInput:
      type: object
      properties:
        percent:
          type: number
          minimum: 1
          maximum: 100
      required:
        - percent